NX-2517: drop assumption that CPUs in /proc/stat are consequtive
Previously, the count of "cpuN" lines in /proc/stat was used to allocate
memory "tables" for per-core metrics. This is not safe since the same
code uses the CPU index (the N from the "cpuN" line) to calculate the
offset in the table. This is because the CPU core indices are not
necessarily sequential or start with zero in /proc/stat. This assumption
is incurrect when some CPUs are explicitly disabled, like with

    echo 0 > /sys/devices/system/cpu/cpu1/online

or when running under RR debugger (RR manipulates the tracee into
believing there's only one core, to reduce the number of threads).

Violating this faulty assumption leads CpuUsageCollector() to perform
many out-of-bounds writes, causing memory corruption and potential
crashes.

This commit makes StartCpuUsageCollector() allocate stats tables based
on the highest CPU core index found in /proc/stat .

We still assume that the "cpuN" lines go in ascending order.

This commit does not address the situation when CpuUsageCollector()
sees newly-enabled CPU cores after the tables were allocated.