NX-2488: enable use of web service cache in requests from NXSL
The code for using cached results was copied from QueryWebService().
This led to a lot of curl API using code to be dropped, for which I am a
bit worried that I might have overlooked some feature buried in it.

This suggests that QueryWebService() and WebServiceCustomRequest() could
benefit from a lot of code reuse.

An automated test was added: run tests/nx-2488/test-websvc-caching on a
fresh disposable install configured to use SQLite.